Viewer refactor
Update and reimport legacy code to fit in new structure
Implemented HashId
Fixed Rust2024 match errors
There's still a bug when zooming in/out on partitions that are not the
origin, hence excess logging in graph_controller.rs. I think it's the
order of operations when you change the anchor partition that's to
blame, but I wanted to get this on the repo before Monday.